Add a program that determines a word's index into a list of its permutations.
This program avoids running in O(n!) time and memory using combinatorics.

Some old code that uses a trie to save space but still suffers from O(n!)
time to generate all permutations is left as a reminder to the author to
think a little harder before starting to code.